requirements.txt
# ── Unified dependencies for multi-transcribe ──────────────────
pyannote.audio
ffmpeg-python
pydantic
rich
pyyaml
mistral-common
numpy
soundfile
transformers>=4.49.0
accelerate
fastsafetensors>=0.1.10
silero-vad
# Whisper now uses transformers (no faster-whisper / CTranslate2 needed)
torchao
bitsandbytes
compressed-tensors
# nemo_toolkit[asr] -- handled in Dockerfile or specialized env

# ── API dependencies ───────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart
pydantic-settings
